Move OnboardingBackground to a sub package
